Fix compilation

+++ pkg/ebpf/tc_probes.go
@@ -90,7 +90,7 @@ func TCIngressDetach(interfaceName string) error {
 	}
 
 	//Currently supports only one handle, in future we might need to cache the handle
-	filterHandle := 0x1 
+	filterHandle := uint32(0x1) 
 	filterParent := uint32(netlink.HANDLE_MIN_INGRESS)
 
 	filters, err := netlink.FilterList(intf, filterParent)
@@ -168,7 +168,7 @@ func TCEgressDetach(interfaceName string) error {
 	}
 
 	//Currently supports only one handle, in future we might need to cache the handle
-	filterHandle := 0x1 
+	filterHandle := uint32(0x1) 
 	filterParent := uint32(netlink.HANDLE_MIN_EGRESS)
 
 	filters, err := netlink.FilterList(intf, filterParent)
